requirements-vertex.txt
# Vertex AI training dependencies (GPU container already has torch/torchaudio)
# torch and torchaudio come from the base pytorch Docker image

soundfile>=0.13,<1.0
numpy>=1.26,<3
scipy>=1.14
pydantic-settings>=2.7
audiomentations>=0.43,<1.0
datasets>=3.0,<4.0
pyarrow>=15.0

# GCP integration
google-cloud-storage>=2.18
google-cloud-aiplatform>=1.72
